+++ packages/article/article.showcase.js
@@ -14,6 +14,7 @@ import fullArticleFixture, {
   longContent,
   videoLeadAsset
 } from "./fixtures/full-article";
+import articleWithKeyFacts from "./fixtures/article-with-key-facts.json";
 import articleAdConfig from "./fixtures/article-ad-config.json";
 
 const preventDefaultedAction = decorateAction =>
@@ -84,6 +85,12 @@ export default {
             leadAsset: videoLeadAsset()
           })
         })
+    },
+    {
+      type: "story",
+      name: "Article with key facts",
+      component: (_, { decorateAction }) =>
+        renderArticle(decorateAction, { fixture: articleWithKeyFacts })
     },
     {
       type: "story",