+++ packages/key-facts/key-facts.stories.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./key-facts.showcase";
+
+showcaseConverter(module, showcase);