+++ packages/key-facts/src/key-facts-prop-types.js
@@ -1,0 +1,8 @@
+import PropTypes from "prop-types";
+
+const propTypes = {
+  ast: PropTypes.object.isRequired,
+  onLinkPress: PropTypes.func.isRequired
+};
+
+export default propTypes;