+++ packages/key-facts/src/key-facts-title-prop-types.js
@@ -1,0 +1,7 @@
+import PropTypes from "prop-types";
+
+const propTypes = {
+  title: PropTypes.string.isRequired
+};
+
+export default propTypes;