+++ packages/key-facts/src/styles/responsive.web.js
@@ -1,0 +1,44 @@
+import { Text, View } from "react-native";
+import withResponsiveStyles, {
+  config
+} from "@times-components/responsive-styles";
+import { spacing } from "@times-components/styleguide";
+
+export const KeyFactsResponsiveContainer = withResponsiveStyles(
+  View,
+  {
+    base: () => `
+    margin: 0 10px;
+  `,
+    mediumUp: () => `
+    flex-direction: row;
+    margin: 0 auto;
+    width: ${config.mediumBpWidth};
+  `,
+    wideUp: () => `
+    width: ${config.wideBpWidth};
+  `
+  },
+  "KeyFactsResponsiveContainer"
+);
+
+export const KeyFactsResponsiveWrapper = withResponsiveStyles(
+  View,
+  {
+    mediumUp: () => `
+    width: 80%;
+  `
+  },
+  "KeyFactsResponsiveWrapper"
+);
+
+export const KeyFactsTitleResponsive = withResponsiveStyles(
+  Text,
+  {
+    mediumUp: () => `
+    padding-right: ${spacing(4)};
+    width: 20%;
+  `
+  },
+  "KeyFactsTitleResponsive"
+);