+++ packages/markup/src/markup.web.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { StyleSheet } from "react-native";
 import Ad from "@times-components/ad";
+import KeyFacts from "@times-components/key-facts";
 import PullQuote from "@times-components/pull-quote";
 import { colours, spacing } from "@times-components/styleguide";
 import renderTreeWithoutDefaults from "./render-tree-without-defaults";
@@ -48,6 +49,12 @@ const defaultRenderers = {
   inline(key, attributes, renderedChildren) {
     return {
       element: <span key={key}>{renderedChildren}</span>
+    };
+  },
+  keyFacts(key, attributes, renderedChildren, indx, node) {
+    return {
+      element: <KeyFacts ast={node} key={key} onLinkPress={() => {}} />,
+      shouldRenderChildren: false
     };
   },
   paragraph(key, attributes, renderedChildren) {