+++ packages/gestures/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.3"></a>
+## [3.0.3](https://github.com/newsuk/times-components/compare/@times-components/gestures@3.0.2...@times-components/gestures@3.0.3) (2018-07-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/gestures
+
 <a name="3.0.2"></a>
 ## [3.0.2](https://github.com/newsuk/times-components/compare/@times-components/gestures@3.0.1...@times-components/gestures@3.0.2) (2018-07-23)