+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.0"></a>
+# [2.1.0](https://github.com/newsuk/times-components/compare/@times-components/styleguide@2.0.2...@times-components/styleguide@2.1.0) (2018-07-23)
+
+
+### Features
+
+* article key facts ([#1122](https://github.com/newsuk/times-components/issues/1122)) ([76035fa](https://github.com/newsuk/times-components/commit/76035fa))
+
+
+
+
 <a name="2.0.2"></a>
 ## [2.0.2](https://github.com/newsuk/times-components/compare/@times-components/styleguide@2.0.1...@times-components/styleguide@2.0.2) (2018-07-23)